+++ NEWS.md
@@ -6,6 +6,14 @@
 * `use_tidy_style()` is deprecated, in favor of using Air for any new, explicit
   formatting efforts (#2110).
 
+* `use_package(min_version = FALSE)` is treated the same as when `min_version`
+  is not specified (#2117, @salim-b).
+
+* usethis's criteria for recognizing a project have expanded to include (#2133):
+  - a `.vscode/` directory, which Positron or VS Code might create
+  - a `_quarto.yml` file, typical of a Quarto project
+  - an `renv.lock` file, which renv creates
+
 # usethis 3.1.0
 
 * `use_vignette()` and `use_article()` support Quarto. The `name` of the new